sp: try lock when executing user initated pool upgrade
When attempting to execute upgradeStoragePool a concurrent
update on one of the pool domains may be during execution and hold
the pool upgrade lock.
In that case, the request to acquire the lock will wait for the lock
to be released which may cause the request to take longer then 3 minutes
and cause to engine failover.

When attempting to upgrade we should attempt to acquire the lock without
wait (try to lock) and if we fail return error immediately, that scenario
is rare and if it happens the user can simply try again.

Change-Id: I93bfdcb19437fdc8901c08ea17dc80fe9c1e5510
Bug-Url: https://bugzilla.redhat.com/1260428
Signed-off-by: Liron Aravot <laravot@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/59514
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-on: https://gerrit.ovirt.org/60520
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Continuous-Integration: Yaniv Bronhaim <ybronhei@redhat.com>