pep8: Excluded files were not excluded from check
The files listed for exclusion were not excluded from PEP8 check. This
had two reasons.

First the exclusion pattern was improperly composed, because $$ was used
to expand variables instead of a single $. This is because the script
snippet originated from Makefile.

Second problem was that `pep8` does accept only one `--except` argument.
This meant that the first pattern list (which contained garbage anyway)
was ignored.

Change-Id: I40f8485f0ced9c117d3416967a11dc6aef3a1dfe
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/59623
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>