v2v: Log detailed output of virt-v2v
The detailed log of virt-v2v output is often necessary to debug
conversion failures. We provide '-v -x' arguments to virt-v2v to get the
detailed output and store the logs in the VDSM log directory.

We cannot use execCmd() to store the log output. Therefore we use CPopen
directly to redirect stderr to stdout and pipe it to tee to store
everything in a log file.

Logs older than 30 days are removed periodically on the daily basis.

Change-Id: I6a8d9284316a551edeaffdd66dfcd299fa02478e
Bug-Url: https://bugzilla.redhat.com/1350465
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/59834
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI