periodic: Avoid useless drive monitoring
The periodic system is checking every 2 seconds if a vm is ready for
commands (Vm.isDomainReadyForCommands). If a vm is ready, it schedules
an extend task in the scheduler (Vm.extendDrivesIfNeeded).

However, if a vm does not have any chunked drives (thin provisioned
block storage), or is in the middle of live storage migration
from non-chunked drive to chunked drive, extending drives is never
needed.

Checking if a vm is ready for commands should be safe, but recently
libvirt fixed a bug that cause this check to get stuck.

Now we check if a vm has chunked drives or non-chunked drives
replicating to chunked drives in DriveWatermarkMonitor.required,
skipping such vms early.

This change saves one libvirt call per vm each 2 seconds, and running
useless extend task per vm in the executor, reducing cpu usage when
using file based storage. This also reduces the risk of uncontrolled
executor worker discarding when NFS server become unresponsive.

Change-Id: I274acc2112cd9b55ac1867d850401be8911adf29
Relates-To: https://bugzilla.redhat.com/1337073
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/59764
Continuous-Integration: Jenkins CI
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>