net test: Nose labeling of test type breaks with inheritance
While running the test suite on Centos7 (with nose 1.3.0), the
functional test module func_net_basic_test disregards the type marking.

The superclass needs to be marked as well.

This patch only fixes the test type marking, the switch type will be
handled in a following patch.

Change-Id: I19bdd61039fdde30fdb30fe6ff477f059f6f85fa
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/59805
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>