migration: usage of single reactor in vdsm
In most of the cases for each standalone client there is reactor but
we want to share one in vdsm. There are already too many threads in vdsm
so it is not desired to add more.

With this change by default client owns a reactor and it will manage its
life cycle. For internal clients we let vdsm to handle its life cycle
because we want to share it.


Backport-To: 4.0
Change-Id: I06e21151e3e9f9e7da9e178bb0199c07f269ae8d
Signed-off-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Bug-Url: https://bugzilla.redhat.com/1349461
Reviewed-on: https://gerrit.ovirt.org/59720
Continuous-Integration: Jenkins CI
Tested-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>