qemuimg.py: Don't check compat support
Modern qemu-img versions allow different QCOW2 formats - 0.10 and 1.1,
although we force the use of 0.10 by default in order to be
backwards compatible with old hupervisors that may still need to
connect to a V3 domain.

Sufficiently old qemu-img versions, though, don't even support the
compat argument, requiring VDSM to query its existence with the "-o ?"
flag before deciding whether to use it. As EL6 support was dropped and
VDSM requires qemu-img>=2.3.0, this check is now redundant, and can
safely be removed, thus cleaning up the code, and saving some CPU
cycles.

As part of this patch, the tests were also cleaned up to match the
code. Tests for qemu-img that does not support the compat argument
were removed, and the remaining tests were renamed to better
remove the notion that having this argument is something special.

Change-Id: Ie141ac33f60b340b7f12832e4fb6b92ea6a68289
Signed-off-by: Allon Mureinik <amureini@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/55925
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>