qemuimg.py: Don't check -T support
Modern qemu-img versions allow using the -T parameter to specify
the cache mode used for the source file(s). In order to be backwards
compatible with older versions (present on EL6, e.g.),
"qemu-img --help" was executed, parsed, and checked for the "-T"
parameter. As EL6 support was dropped and VDSM requires
qemu-img>=2.3.0, this check is now redundant, and can safely be
removed, thus cleaning up the code, and saving some CPU cycles.

As part of this patch, the tests were revised to include this
argument, instead of just mocking qemu-img that does not support it.
The tests for the checking mechanism were removed completely,
together with this mechanism.

Change-Id: Ic4ac5f44b2788b76a49440330033dfc5faea0831
Signed-off-by: Allon Mureinik <amureini@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/55926
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Continuous-Integration: Jenkins CI