spec: Require lvm version fixing filter
Older lvm version was using filter incorrectly when running commands
such as vgck and pvs. Instead of rejecting all the devices in the
filter, the commands opened and scanned all block devices. This slow
down these commands on systems with lot of block devices.

On 3.6, lvm slowdowns cause delays in storage monitoring and may cause a
host to become non-operational. These delays may also cause storage
operations such as copying disks to fail with a timeout.

On 4.0, these slowdowns may cause unwanted delays in various flows, but
are not harmful to storage monitoring.

Fedora 23 and later already include this fix (lvm2-2.02.132-2.fc23).

Change-Id: I24040df9e19b4a093ce7e4f954375017f694099f
Bug-Url: https://bugzilla.redhat.com/1346012
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/59907
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>