net ifcfg: fix bond rollback
When setupNetworks(NET, BOND) fails on NET configuration, BOND ifcfg
file is removed, but it still exists in the system. We have to remove
it manually like we do in removeBonding().

This must be backported to fix new functional tests on 4.0.

Change-Id: I5cea6ec71c913d74d95317ff7318259d64b40969
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/59940
Reviewed-by: Edward Haas <edwardh@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1374194
Reviewed-on: https://gerrit.ovirt.org/64509
Reviewed-on: https://gerrit.ovirt.org/65303
Reviewed-by: Francesco Romani <fromani@redhat.com>
Tested-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Francesco Romani <fromani@redhat.com>