vdscli: Replace the default '::' host with hostname
vdscli may fail to connect to the VDSM server when using the default
configuration settings of vdsm.conf.
By default, in vdsm.conf, the management host address is set as '::',
but while the server can fallback and listen on IPv4 addresses if the
IPv6 stack is unavailable, the client does not, therefore it fails to
connect to the server.

This patch adds client logic that replaces a '::' host with the system
hostname. In turn the hostname will get resolved to the available address.

Change-Id: Iff0401767849129a7075a57e8cb3ff6af2b2d060
Bug-Url: https://bugzilla.redhat.com/1350883
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/60020
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
(cherry picked from commit 5e5431207b42103c999b232f8b93e7d0d9bddcd9)
Reviewed-on: https://gerrit.ovirt.org/61363