qemuimg: Unbreak qmemuimg.check
Since commit 3e7a11954582 (qemuimg: Use --output json for 'check' and
'info') qemuimg.check is broken, but the tests are passing. This is
caused by unneeded mocking.

qemuimg.check fails now with this error:

      File "/usr/lib/python2.7/site-packages/vdsm/qemuimg.py", line 309, in _parse_qemuimg_json
	obj = json.loads(output)
      File "/usr/lib64/python2.7/json/__init__.py", line 338, in loads
	return _default_decoder.decode(s)
      File "/usr/lib64/python2.7/json/decoder.py", line 365, in decode
	obj, end = self.raw_decode(s, idx=_w(s, 0).end())
    TypeError: expected string or buffer

Replace the harmful mocking with running qemuimg.check with a real
image, and add the missing raw=True to get a string from the command.

Change-Id: I732b8bb7285352ed92d69b295b8cf6c2598b8f41
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/60168
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>