tests: Terminate watcher when tests fail
When the tests fail, the watcher process was not terminated. It would
continue to run for 300 seconds in the background, and then try to kill
the runner process, possibly killing unwanted process.

This happens only when running tests locally; on the CI the watcher was
terminated by chroot.

Change-Id: Ief975ac59c8c3ec106c42c93d5cb2a2fd32b69c1
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/60239
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>