vm: graphics: report display info from device
We used to keep the first graphic device information in Vm.conf,
to keep backward compatibility with ancient (< 3.6) Engines.

But now we are allowed to break it, hence we will get rid
of the duplicated data in Vm.conf.
But the API schema still require Vdsm to provide this information
in the stats output.

So, this patch makes the code fetch the needed info from the device
info, no longer from the generic Vm.conf.

Change-Id: I1f501addd40458887ae27cd9615a8856a02d8967
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/60320
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI