clusterlock: Support for multiple leases
Previously the clusterlock api was supporting only a single lease (the
spm lease). The lease path was set when creating a cluster lock, and used
with hardcoded offset and name when acquiring, releasing and inquiring
the lease.

The leases path was removed from cluster lock api. Instead, initLock(),
acquire(), inquire() and release() receive now a lease argument, which
is a named tuple (cluserlock.Lease) with the lease details.

The details of the spm lease are kept now in sd.py, passing the lease to
the cluster lock.

This patch is based on Adam's patch https://gerrit.ovirt.org/38622. But
instead of introducing new clusterlock methods to work with volume
leases, the existing api is modified to accept a lease.

Change-Id: I7903b13bd017e3da6762e8979f7f4dfc6520de9e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/60422
Reviewed-by: Adam Litke <alitke@redhat.com>