vm: api: move vm status check into migrate()
The API layer is guarding against incorrect calls of
the VM.migrate(). This method should not be called if the
VM is in WAIT_FOR_LAUNCH or DOWN statuses, two unlikely conditions
anyway.

To have this protection, we need to import vmstatus into API.py,
and we tight the coupling between VM and API.

Turns out VM.migrate() already partially does the very same safety
check inside its implementation.
This patch fully move the check which API used to do inside
VM.migrate(), to loosen the coupling and improve isolation.

Change-Id: Ic7d5bfcc98b1169e2deed947c945e55c19d5ab04
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/60608
Continuous-Integration: Jenkins CI
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>