qemuimg: Move wait_for_completion inside QemuImgOperation
The Image class implemented logic to wait for a QemuImgOperation to
complete.  This same logic is useful to SDM.copy_data and should be
shared.  Move it inside QemuImgOperation and hide the old wait function
which no longer needs to be public (renamed to _poll).

Change-Id: I95075798022e67b0d1e63e7f7677ba0f84c2b328
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/60691
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Nir Soffer <nsoffer@redhat.com>