storage: add an option to zero volumes using "blkdiscard --zeroout"
Up until now, we used "dd" to zero volumes.
This patch introduces a new method to zero volumes, using
"blkdiscard --zeroout".
Note that this patch keeps the old "dd" method as the default one for
zeroing volumes, and that it can be configured by setting the config
value "zero_method".

When "blkdiscard --zeroout" is used:
1. If the underlying storage supports "write same" [1], then the kernel
   quickly zeroes the volume using write same.
   This can take ~ 90% less time than zeroing it using "dd".
2. Otherwise, the kernel zeroes it by writing pages of zeros, just like
   when using "dd".

[1] For more info, please refer to the feature page "Wipe volumes using
blkdiscard":
http://www.ovirt.org/develop/release-management/features/storage/wipe-volumes-using-blkdiscard/

Change-Id: I40bb6aa3d431e48aa87af899a6b9b61d000bff1a
Bug-Url: https://bugzilla.redhat.com/1367806
Signed-off-by: Idan Shaby <ishaby@redhat.com>