storagetestlib: introduce the Aborting tests tool
By using the Aborting tests tool, one can create an object that when
called for the first N times it returns False, and from the N+1 call and
on it returns True.

It is handy when testing tools that use commands.watchCmd, that can wait
for a stop signal and then kill the execution of the new process.

Here's an example of a test to assert that cmd aborts on the third poll:
with self.assertRaises(ActionStopped):
    commands.watchCmd(cmd, stop=Aborting(2))

Change-Id: I83e6e60e0e3e62455700eddabc36c5a372059fc6
Related-To: https://bugzilla.redhat.com/1367806
Signed-off-by: Idan Shaby <ishaby@redhat.com>