net: Silently skip persisting a non-existent running-config
In vdsm-store-net-config unified, running-config is persisted by
copying it to a dedicated location.
If for some edge case scenario, the running-config is absent, an error
message appears in the log, without failing the script run. (cp fails)

The scenario in which this occurs is limited to an initial state
where there are no VDSM owned networks, therefore, no running-config
file is created.

Change-Id: If95795f79872d0ae404c4fc4b3be1babac1e2d24
Bug-Url: https://bugzilla.redhat.com/1329387
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/60744
Continuous-Integration: Jenkins CI
Reviewed-by: Petr Horáček <phoracek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>