oop: Explicitly close running ioprocesses
IOProcess client is using __del__ for automatic closing when the last
reference to the client disappears. This attempt fail consistency when
__del__ is called during shutdown, since Python is partly tore down.

In the worst cases, the entire application will get stuck during
shutdown. We see this randomly when running the tests in the CI.
Looking in backtraces, all the threads are blocked trying to acquire a
lock that may have been already destroyed.

This patch avoid this issue by explicitly closing all running
ioprocesses during application shutdown and when cleaning up after
storage tests. This ensure that IOProcess's close() and __del__() are
running in a well defined environment.

Change-Id: Id9fce2bcc8a7916dd9e31a8b23a4199611c7938f
Bug-Url: https://bugzilla.redhat.com/1334274
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/60830
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>