spec: Require qemu-kvm-ev to fix aio_bh_poll hang
This qemu-kvm-ev build fixes a hang experienced during live merge
(original reported in the Related-To BZ, and since cloned several
times to the various platforms/releases).

This patch bumps vdsm's requirement for qemu-kvm-ev to deliver this
fix and ensure oVirt users get the smoothest experience when live
merging a snapshot.

Although the issue technically just requires bumping the qemu-kvm-ev
requirement, the qemu-img-ev requirement was bumped together with it,
in order to preserve consistency and avoid mixing versions by mistake.

Change-Id: I29639a2de15cb325a7aa11ab9fcd23ab1359b2d7
Related-To: https://bugzilla.redhat.com/1319400
Bug-Url: https://bugzilla.redhat.com/1349529
Signed-off-by: Allon Mureinik <amureini@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/60859
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>