tests: Set vars/ssl to true in graphics tests
In 8a7479c, support for graphics defaultMode was added and
defaultMode="secure" is now expected in test outputs.  However this is
only present when vars/ssl configuration option is set to true.  We must
ensure this is so by monkey patching the corresponding graphics tests.
Otherwise the configuration value from /etc/vdsm/vdsm.conf (if present)
is apparently used and the tests may fail if the value is false.

Change-Id: I31df7639f23f9cf813d0baf1bb102920bea76736
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/60973
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Tested-by: Francesco Romani <fromani@redhat.com>