virt: Introduce MutableDomainDescriptor
DomainDescriptor is a useful utility for handling domain XML and we
would like to use it more.  An important property of DomainDescriptor is
that it assumes to be used read only and so it caches some of its data
and doesn't expect changes to them.

In Vm._correctDiskVolumes, we process domain XML "manually".  But we
already have the corresponding functionality available in
DomainDescriptor.  The only problem is we want to modify the parsed DOM
in place.  For this purpose, we introduce a new class
MutableDomainDescriptor in the form of a mutable base class for
DomainDescriptor.  The class is going to be used in the followup patch.

Change-Id: I9e54129472da12f49d72ae96eadf3c7c09c55148
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61105
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI