hostdev: add back setup log message
Lost in the refactoring process, although the importance of the message
itself is debatable, it does provide nice indication of state change
and consistency between host and network devices.

The message is only displayed for PCI device as other devices are not
really detached from the host (they piggyback on detach_detachable's
capability detection code).

Change-Id: I5818fcff3c930adfe7a7c5cd0d97ecfcd4a51058
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61181
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>