lvm: Refresh lvs during activation
Previously we used to activate only the inactive lvs, and ignore the
active lvs silently. An active lv may not reflect the mapping on
storage if the lv was extended or removed on another host.

In case of incorrect size, starting a vm with an lv with incorrect size
may cause the vm to pause when trying to write to the lv. This may lead
to data corruption when migrating a vm from another host, when the lv
size on the source machine is bigger then the lv size on the destination
host.

In the case of removed lv, writing to the lv may corrupt an extend
belonging now to another lv. This should never happen since we do not
use lvs after we remove them.

To avoid these issues we refresh active lvs during activation.

Change-Id: I845ce1f930d7673ec16cfa0a11399081ae2e6caa
Bug-Url: https://bugzilla.redhat.com/1358348
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61285
Continuous-Integration: Jenkins CI
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/62731
Tested-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>