lvm: Improve logging lv activation logging
Log vg name and lv names when we activate, refresh or deactivate lvs.
All are important state changes that should be logged in INFO level.

The actual lvm commands are logged in DEBUG level, and may not be
available when we change the default log level to INFO.

Change-Id: I368b158c8d60ae2ff252e038b01024e4598b5f4d
Bug-Url: https://bugzilla.redhat.com/1358348
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61286
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Continuous-Integration: Jenkins CI