lvm: Never hide lvm errors
In changelv, if lvm command has failed but stdout is not empty, the
error was ignored silently. This may hide errors when deactivating lvs
leading to stale lvs, or hide other errors.

This code was introduced in 2009 with the first version of the lvm
module, and there is no documentation why ignoring errors when stdout is
not empty is needed.

Testing lvm show that using verbose mode (-v) will write to stdout and
cause errors to be hidden. I could not find lvm command that requires
this strange treatment, and no other lvm command is using it.

This patch removes the stdout empty check, raising if lvchange return
non-zero exit code.

Change-Id: I6ce160ecddebfc903fb7bc00ba5a69e63f2cc996
Bug-Url: https://bugzilla.redhat.com/1358348
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61287
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Continuous-Integration: Jenkins CI