compat: py3: capture output of CPopen.communicate
To allow better coverage of python3 testing, we use Popen instead of
CPopen, as the latter is not implemented (and not needed) in python3.

Unfortunately, CPopen has chosen to use stdout\err=subprocess.PIPE as its
default. Now when cpopen supports stdout and strerr params we can always set
those to PIPE to match the defaults.

Change-Id: Iac9c1c1bd3ca1902103fac1f8675b7d42ca78a33
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61393
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Irit Goihman <igoihman@redhat.com>