vm: periodic: fix stats age reporting
We need to report the Vm responsiveness, because it is one
important indicator of the Vm health. This information can
trigger the administator intervention, or automatic corrective
action (e.g. Vm migrations).

The Vm class reports the monitor responsiveness in two ways:
1. if a libvirt operation reports a timeout (a standard libvirt
exception)
2. if one sampling operation hangs for too long

In the second case, given the transient nature of the detected
failure, the responsiveness is reported, but not stored as Vm
field. This is noteworthy, but totally expected and legitimate.

The sampling code stores the bulk samples only for a limited
number of cycles, and drops the old values.
The sampling code also used to calculate the sample age
based on the stored values.

But if a VM hangs for too long, all its samples are eventually
evicted from the samples cache, so its stats age becomes None,
which is misinterpred as responsive VM.

We should keep reporting None when we don't have enough meaningful
data, but we should also fix the sampling age reporting;
To do so, we leverage the last recorded vm timestamp.
This timestamp is needed by the flow, so it is guaranteed to be present.

This patch leaves two potential races open. We do so because we
think both are insignificant:

1. the sampling stats cache is initialized in Vm._domDepedentInit(),
so if a VM takes too long (> vars.vm_command_timeout, default 60s)
to get created, it will be reported as unresponsive until the domain
is ready. This is actually a more correct representation of reality.
2. if the powerdown flow takes too long (same as #1), again the
VM will be reported as unresponsive until the domain is completely
cleaned up. This is also a more thrutful reporting.

Change-Id: I3e61626625a2e0517d55dc61e361f3f5eb690c00
Bug-Url: https://bugzilla.redhat.com/1360986
Backport-To: 4.0
Backport-To: 3.6
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61310
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61822
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/62418
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>