ovs: create ovs-vsctl command only on demand
If openvswitch package is not installed, import of ovs modules fails
because of ovs-vsctl command cannot be created.

With this patch, the command was moved to memoized function and is
evaluated only on demand.

Change-Id: If903a1452254a8776b91fb7f11a13eabbcbb8c80
Bug-Url: https://bugzilla.redhat.com/1195208
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61321
Reviewed-by: Edward Haas <edwardh@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
(cherry picked from commit d8a3fcb5616ddb18dc3cdc360ce0fcbeb961c330)
Reviewed-on: https://gerrit.ovirt.org/61347
Tested-by: Edward Haas <edwardh@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>