modprobe: set bonding max_bonds to 0
When bonding kernel module is loaded, bond0 is implicitly created.
When we create a bonding during installation of the system in
Anaconda, it obtains the same name bond0.
On the following boot, NM fails because of conflict between NM handled
and implicitly created bond.

This patch introduces modprobe configuration file for bonding module.
With `options bonding max_bonds=0` we disable implicit creation of
bondings.

Bug-Url: https://bugzilla.redhat.com/1356635
Change-Id: Ic264e8a89eb07b8da42494a9b12ac9c871f78c51
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61396
Continuous-Integration: Jenkins CI
Reviewed-by: Fabian Deutsch <fabiand@redhat.com>
Tested-by: Fabian Deutsch <fabiand@redhat.com>
Tested-by: Edward Haas <edwardh@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>