virt: sampling: add is_empty() method to StatsSample
One upcoming patch wants to fill the 'stats_age'
field with meaningful value.
The problem is that None is misinterpreted by the Vm code,
so if the stats are too old or missing, the Vm is not
detected as unresponsive.

So, the EMPTY_SAMPLE is not good anymore, and this patch
removes it.
We still need to know if a given sample is empty, meaning
it doesn't hold meaningful data, so we add one is_empty()
method to fulfill this task.

Another option could be implement __nonzero__, but it
doesn't completely fit in this case.

Change-Id: I7bc3712d36c46ade9a779411028817c197ec34b3
Bug-Url: https://bugzilla.redhat.com/1360986
Backport-To: 4.0
Backport-To: 3.6
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61420
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61821
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/62417
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>