storage: Introduce guarded utilities
Throughout the storage code we have implicit locking rules which we
follow in order to ensure that concurrent operations will not conflict
with one another.  We also require a specific locking order to prevent
deadlocks.  This code is repeated throughout the storage subsystem and
is error prone.  To alleviate this, introduce the guarded module with a
context manager that can perform the necessary locking.  At first we
will support resourceManager locks and volume leases.

The contextmanager accepts a variable number of lock lists where each
list corresponds to the locks required to protect one element of the
operation (eg.  a source volume).  Lists must contain only AbstractLock
derived objects.  All locks from all lists are aggregated and locked
according to the rules.  When exiting the context (or upon error) the
locks are released in reverse order.  If an exception was raised inside
the context, that exception will be propagated regardless of any lock
release failures that may occur.

The following rules are currently enforced:
- Locks are collected from all lists and sorted using namespace,
  name, and mode (in that order)
    - Due to namespace sorting, ResourceManager locks are taken before
      volume leases
- Duplicate locks are removed (only when ns, name, and mode match)
- Locks are acquired in sorted order and released in reverse order

Change-Id: I2b0a204818d44b6205515277f4c2834cb2b7a057
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61435
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI