yajsonrpc: use proper exception in the recovery
If recovery takes too long and if we get an API call meantime,
the yajsonrpc correctly tries to reply with "recovery in progress"
error.

The problem is that in this flow - and looks like only in this flow -,
we raise the JsonRpcError with a response object, not with a proper
exception object - as the TODO very next to the changed code
reminds us.

So, we get

jsonrpc.Executor/1::ERROR::2016-07-28
12:52:25,966::executor::290::Executor::(_execute_task) Unhandled
exception in Task(callable=<functools.partial object at 0x1c71db8>,
timeout=None)
Traceback (most recent call last):
  File "/usr/lib/python2.7/site-packages/vdsm/executor.py", line 288, in
_execute_task
    task.callable()
  File "/usr/lib/python2.7/site-packages/yajsonrpc/__init__.py", line
512, in _serveRequest
    response_log = "failed (error %s)" % (error.code,)
AttributeError: 'dict' object has no attribute 'code'

This issue was present since quite some time, but we recently merged
78bdaa1d6578b6491d77fa9432f555fe05c1c0c2 that exposed it.
There is no evidence this happens outside the recovery flow, and the
error is not fatal, the only impact is a nasty stacktrace in the logs
and an unexpected failure returned to clients.

Change-Id: Ic839e53d8fbbb4ee8e231153e30b3d6be30f9b00
Backport-To: 4.0
Backport-To: 3.6
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61613
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>