virt: Fix of Executor._active_workers crash on modification
Executor._active_workers doesn't use a lock when accessing
self._workers.  This is intentional, since we access self._workers
read-only, we don't need exact value for some snapshot of the executor
state and we have no reason to impose locking on the method.

However, we still must be careful, because we loop over the value.  If
it gets modified during the loop then

  RuntimeError: Set changed size during iteration

is raised.  This patch prevents that error by making a shallow copy of
self._workers in the loop.

Change-Id: Iaed995109e98c5ab79724a4765c3212140ebebb7
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61690
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>