net: Introduce nameservers (dns) network api
oVirt is required to preserve and manage the nameservers for the host.

This patch introduces the api layer through which nameservers may be
set.
Nameservers can be set through setupNetworks verb, using the 'nameservers'
key in the networks sub-verb where the value is a list of nameservers.

Canonicalization: In case the key is missing, it will be populated with
the current dns setting (from resolv.conf) for the management network
(identified by the defaultRoute flag).

Change-Id: Ic9b0d0370951fe3985f61e550374eb63eb2cf161
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61739
Continuous-Integration: Jenkins CI
Tested-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1351095
Reviewed-on: https://gerrit.ovirt.org/61931