functional tests: fix supervdsmServer test
supervdsmServer dropPrivileges function is switching root user to
a non privileged user but not switching the user back to root. It causes
other tests which require root user to fail.
this patch runs the tests in a child process so there's no affect on the
parent process.

Change-Id: I8c7e381809261cf330ef4f19c3fec5cd38254b1b
Signed-off-by: Irit Goihman <igoihman@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61902
Continuous-Integration: Jenkins CI
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>