jsonrpc: shorten executor thread system names
We shorten the basename of the jsonrpc executor,
because the system has a hard limit of 15 ASCII
character per thread name.

The new name is less expressive, but leaves room
for the worker ID, which is an overall usability
improvement.

Before, it looked like

  jsonrpc.Executo

now it looks like

  jsonrpc/0

Change-Id: Ibefc7d2c60d7b5d23739b129088f8e7f80514540
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61978
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Irit Goihman <igoihman@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>