automation: build also vhostmd hook
In 3.6 we shipped vdsm-hook-vhostmd as part of our VDSM releases.
In 4.0 this hook is missing from the build due to a bug in the
automation code which didn't configure the spec file for creating
the sub-package.

This patch address this regression, re-adding vdsm-hook-vhostmd to
the packages shipped within ovirt repos.

Change-Id: I249131cf4d3a3623a32708b961a5273fba4a69eb
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/61984
Continuous-Integration: Jenkins CI
Reviewed-by: Gil Shinar <gshinar@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
(cherry picked from commit b9a2e7306e286708998d9035ffa08ffc9f26737c)
Reviewed-on: https://gerrit.ovirt.org/62268