jobs: Move run from sdm.base.Job to jobs.Job
The run implementation in sdm.base.Job is generally useful.  It performs
state checking and error handling that can be applicable to all job
types.  Move it to job.Job.  The tests associated with this
functionality need to move to jobsTests.py which means we no longer need
storage_sdm_api_test.py or sdmtestlib.py

Change-Id: I194eb154d70e846b7ec5d462faa26b3aa35d2ff6
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/62001
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI