Removed %config(noreplace) from %{_tmpfilesdir}/%{vdsm_name}.conf
%{_tmpfilesdir} expands to %{_prefix}/lib/tmpfiles.d which is the
location that the package's default tmpfile creation scripts should
install into. %{_tmpfilesdir}/%{name}.conf is not marked as a %config
file because it is not supposed to be edited by administrators.
Administrators can override the package's %{name}.conf by placing an
identically named file in /etc/tmpfiles.d/, but this should very rarely
be needed.

Change-Id: Ibf3268da3e601083c6c67771878da82be66da341
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/62030
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Continuous-Integration: Jenkins CI