sdm: Drop support for thread-local job_id
sdm.api.base.Job currently has support for setting the job_id in the
thread-local variables.  We are not using this feature so let's remove
it and simplify our _run code and tests.  This will also _run (and
associated tests) to move to the general vdsm.jobs.Job class without
dragging in storage-specific interfaces.

Change-Id: Ia4697f2d3f5e6826e0e95e03a10445c2df2abddf
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/62083
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI