net: don't accept nameservers on a non-default network
Currently, only the default/management network (the one with
the default route) may be used to set nameservers, for reasons
given below. Previously, an attempt to set nameservers through
a non-default network would be ignored silently.

While it may be possible in the future to refine this design,
the configuration of nameservers is now viewed as a per-host
task, in which it makes sense to use the default network.

Change-Id: Ib42a5d5a7aeb1169d510a3de8d04645efbab505e
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/62167
Reviewed-by: Edward Haas <edwardh@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1351095
Reviewed-on: https://gerrit.ovirt.org/62362