api: emit warnings only if configured
The python standard module `warnings` implement
filtering for warnings, to allow the client choose
how often the warning should be presented, like every time,
or the first time only and so forth.

However, to implement this behaviour the module silently
adds the __warningregistry__ dicts to various modules;
furthermore, the key of the entries include the message
of the warning being checked.

Vdsm uses the warnings module to report inconsistencies
between the schema and the actual data received or sent;
we check the responses, and this includes variable data
(e.g. in any stats reporting flow), including timestamps.

This means that every warning we issue will have a different
string representation, thus a different entry in the
hidden __warningregistry__, with the ultimate result of
unbound growth and unbound resource usage.

One possible workaround is to actually enable the warning reporting,
setting

  [devel]
  python_warnings_enable = true

in /etc/vdsm/vdsm.conf.

This makes the warnings module skip the usage of its registries,
sidestepping the issue. The problem, however, is that logs are
much more verboses

This patch implements another mitigation: we now issue warnings
only if requested - using the same configuration item.
This prevents entirely the warnings to be raised in this flow.

Please note that this patch is not definitive, we still need
better handling of warnings project-wise.

Change-Id: If17560df93fa6cb58e86eedaca62b893a85ed98b
Bug-Url: https://bugzilla.redhat.com/1364149
Backport-To: 4.0
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/62186
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>